Finish refactoring API
