+++ web/src/routes/+page.ts
@@ -1,7 +1,7 @@
 import type { PageLoad } from "./$types";
 
 export const load: PageLoad = async ({ fetch }) => {
-  const r = await fetch("api/models");
+  const r = await fetch("/api/model/");
   const models = (await r.json()) as string[];
   return {
     models,
